Fix orgit-file-store to add links to org-stored-links when interactive
* orgit-file.el (orgit-file-store): Call `org-link--add-to-stored-links'
when called interactively to make stored links available in
`org-insert-link'.  Previously, the function only set
`org-store-link-plist', which is sufficient when called through
`org-store-link' (which handles adding to `org-stored-links' itself),
but insufficient when called directly.  Update docstring to document
that interactive calls add links to `org-stored-links'.

This fixes the issue where calling `orgit-file-store' interactively
would display "Stored: ..." message but the link wouldn't appear in
`org-insert-link' or `org-stored-links'.